requirements.txt
###### Requirements without Version Specifiers ######`
folium
pandas
requests
pyproj
geopandas
urllib